Version 24.3.4
[Engine]
* CppCheck reported ambiguity with the use of 'operator bool(void)' in both the 'FileMap' class and the 'FileMap::SysMap::FStream' (POSIX) class and was resolved by moving everything from 'FStream' into a new 'FStreamBase' class without the 'operator bool(void)' function, moving that function in to the 'FStream' class (which inherits the FStreamBase class) and switching 'FStream' to 'FStreamBase' in the 'SysMap' (POSIX) class.
* Fix other minor CppCheck recommendations.
* Move 'FStream::Mode' out of its class to the 'IFStream::P' namespace for easier access and update all references to it.
* Remove obsolete 'SysInitThread' from pixbase.
* Rewrite 'SysInitThread' to properly set thread priority on MacOS, and disabled setting thread priority completely on Linux since it requires elevated privileges to do, Windows code being simplified and SysThread::* demoted to a regular 'enum' so it can be 'size_t' type.
* Fixed wrong uptime being reported on Linux because STL hires clock class isn't counting from zero when the system starts.
* Fixed graphical corruption when window threading enabled on Wayland because of window size not being initialised immediately on creation.
* Hurray! 'win_thread' can now default to 1 on ALL systems.
* Disable input when using the FS toggler key function to prevent those keys (or any input) being repeated on Linux. It's still not perfect but helps until Ubuntu updates their GLFW because 3.4.0 fixes it.
* Workaround for using the FS toggler key to go back to desktop window mode not working properly on Linux by just hard resetting the window .
* Fix potential nullptr reads on demangle error checking.
* Remove unnecessary extra parameter on image RGB re-order message.
* Use 'cosf/sinf/atanf/sqrtf' functions in 'SetVertex' function since only working on 32-bit floats.
* Don't force 10bit framebuffer on Linux by default to prevent initialisation error.
* Use GLFW_DONT_CARE for alpha bits.

[Assets]
* Update Diggers readme to reflect fixes.

+++ src/fboitem.hpp
@@ -81,19 +81,19 @@ class FboItem
   /* -- Set vertex bounds with angle --------------------------------------- */
   void SetVertex(const GLfloat fX1, const GLfloat fY1,
      const GLfloat fX2, const GLfloat fY2, const GLfloat fA)
-  { // UtilDenormalise the angle to radians                             (M_PI)
+  { // UtilDenormalise the angle to radians                          (M_PI)
     const GLfloat fAR = fA * 2.0f * 3.141592653589793238462643383279502884f,
     // Calculate centre
     fXC = (fX2-fX1)/2, fYC = (fY2-fY1)/2,
     // Rotate vertices
-    fC1 = atan2(-fYC,fXC)+fAR,         fC2 = atan2(-fYC,-fXC)+fAR,
-    fC3 = atan2( fYC,fXC)+fAR,         fC4 = atan2( fYC,-fXC)+fAR,
-    fC5 = sqrt(-fYC*-fYC+fXC*fXC),     fC6 = sqrt(-fYC*-fYC+-fXC*-fXC),
-    fC7 = sqrt( fYC* fYC+fXC*fXC),     fC8 = sqrt( fYC* fYC+-fXC*-fXC),
-    fCa = cos(fC2)*fC5,                fCb = sin(fC2)*fC5,
-    fCc = cos(fC1)*fC6,                fCd = sin(fC1)*fC6,
-    fCe = cos(fC4)*fC7,                fCf = sin(fC4)*fC7,
-    fCg = cos(fC3)*fC8,                fCh = sin(fC3)*fC8;
+    fC1 = atan2f(-fYC,fXC)+fAR,         fC2 = atan2f(-fYC,-fXC)+fAR,
+    fC3 = atan2f( fYC,fXC)+fAR,         fC4 = atan2f( fYC,-fXC)+fAR,
+    fC5 = sqrtf(-fYC*-fYC+fXC*fXC),     fC6 = sqrtf(-fYC*-fYC+-fXC*-fXC),
+    fC7 = sqrtf( fYC* fYC+fXC*fXC),     fC8 = sqrtf( fYC* fYC+-fXC*-fXC),
+    fCa = cosf(fC2)*fC5,                fCb = sinf(fC2)*fC5,
+    fCc = cosf(fC1)*fC6,                fCd = sinf(fC1)*fC6,
+    fCe = cosf(fC4)*fC7,                fCf = sinf(fC4)*fC7,
+    fCg = cosf(fC3)*fC8,                fCh = sinf(fC3)*fC8;
     // Update the first triangle of the quad
     TriPosData &tdT1 = GetVData(0);
     tdT1[0] = fX1+fCa; tdT1[1] = fY1+fCb; // Vertex 1 / Triangle 1
